Fix issue where multiple workers get the same unique_suffix
From Random documentation:
The functions supplied by this module are actually bound methods of a hidden instance of the random.Random class. You can instantiate your own instances of Random to get generators that don’t share state. This is especially useful for multi-threaded programs, creating a different instance of Random for each thread,

we used to use random.choices from two threads. It would call the same
method on a none thread safe implementation of random and it would
return the same value when two threads spawn at the same time and call
the same implementation.

The new change creates per thread an implementation seeded by the system
clock. This is not a StrongRandom like SystemRandom but it is not OS
dependant and should be good enough.